Move MessageLocalization and TypeNameLocalization to helper package
