+++ src/main/java/net/sourceforge/argparse4j/helper/MessageLocalization.java
@@ -1,4 +1,4 @@
-package net.sourceforge.argparse4j.internal;
+package net.sourceforge.argparse4j.helper;
 
 import net.sourceforge.argparse4j.inf.ArgumentParser;
 import net.sourceforge.argparse4j.inf.ConfiguredArgumentParser;
