+++ src/test/resources/net/sourceforge/argparse4j/helper/TypeNameLocalizationTest$CustomTypeWithResourceBundle-argparse4j.properties
